+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - dial: use underlying net.Dialer's DNS Resolver in Dialer.
 - stanza: change API of `WrapIQ` and `WrapPresence` to not abuse pointers
 - xmpp: add new `SendIQ` API and remove response from `Send` and `SendElement`
+- xmpp: new API for writing custom tokens to a session
 
 ### Fixed